social_media_api/requirements.txt
Django>=5.2.7
djangorestframework>=3.16.1
Pillow>=12.0.0
